Only initialise the PTM connection when running a node instance, instead of a subcommand.
